app/model.py
import random
import io
from PIL import Image
from app.schemas import RecommendationItem

class FashionRecommender:
    def __init__(self):
        # Here we would load actual model weights, e.g. using PyTorch or TensorFlow
        self.categories = ["Jacket", "T-Shirt", "Jeans", "Sneakers", "Hat"]
        self.colors = ["Black", "White", "Navy", "Beige", "Red"]

    def predict(self, image_bytes: bytes) -> list[RecommendationItem]:
        \"\"\"
        Process the image and return recommendations.
        \"\"\"
        # Convert bytes to PIL Image just to verify that it's a valid image
        try:
            image = Image.open(io.BytesIO(image_bytes))
            # Just do something with it to prevent dead code warnings, like accessing size
            _ = image.size 
        except Exception as e:
            raise ValueError("Invalid image file provided.")

        # Simulate inference by picking random categories and colors
        num_items = random.randint(1, 3)
        recommendations = []
        for _ in range(num_items):
            recommendations.append(RecommendationItem(
                category=random.choice(self.categories),
                confidence=round(random.uniform(0.6, 0.99), 2),
                color=random.choice(self.colors),
                description="Our model thinks this matches well."
            ))
            
        return recommendations
