requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
Pillow==10.2.0
pydantic==2.6.4
